added first draft of multi-column aggregation

+++ src/main/java/org/whiskeysierra/helsing/hadoop/functions/CorrelationCoefficient.java
@@ -26,7 +26,7 @@ final class CorrelationCoefficient extends TwoArgumentNumericAggregator {
     }
 
     @Override
-    public Long getResult() {
+    public Double getResult() {
         return covariance.getResult() / (leftDeviation.getResult() * rightDeviation.getResult());
     }
